Summary of What I have so far
Using SDL libraries, the Laws of Physics and with a little patience and inspiration from Box2D. So far, I have made an engine with some convienet functionality in regards to Objects.

Current Goal:
Make the engine more interactive through the window and the concsole as well. Taking the user input and feedback and the ability to create objects inside the SDL Window as well.