End of Semester 1 Push
Did a lot of good work to the engine this past semester. The Body Defintion of the Body objects take inspiration from Box2D where upon instantiation it will have default values, you may change the values as you please. Added Key board input as well so that each scene inherits an Input Manager in order to handle the user's inputs. I a m still researching some more dynamic ways of doing this, but it does the job as of right now.

Going Forward I will be encorporating circular motion and collisions with the intention of providing the user with as much control as posibble over the physics and general events in the systems of the engine.

Some more things I looking forward to implementing are:
-grpahics
-sprites/ texture redering
-more to come..