High-Resolution Timer, Debug Log & SwapChain SetUp
Using Performance queries I created a high resolution timer, I will be able to use this timer(with some added functionality) when it comes to profiling stats with the engine. I also added a debug static class with functions to record errors, info and warnings for the engine. I will be growing this class as needed during development. Swap chain set up is now complete. Next up is setting up the grpahics pipeline for vulkan